Funcionalidad de boton eliminar

+++ src/lib/firebase.js
@@ -2,7 +2,7 @@
 import { initializeApp } from 'firebase/app';
 import { getAuth, createUserWithEmailAndPassword, GoogleAuthProvider, signInWithPopup } from 'firebase/auth';
 import { firebaseConfig } from './firebase-config.js';
-import { getFirestore, collection, addDoc, getDocs, orderBy, query, deleteDoc, doc, setDoc } from 'firebase/firestore';
+import { getFirestore, collection, addDoc, orderBy, query, deleteDoc, doc, onSnapshot, updateDoc } from 'firebase/firestore';
 
 
 // Initialize Firebase
@@ -37,46 +37,65 @@ export const loginGoogle = () => {
 // guardar post
 export const guardarPost = (post) => {
   const user = auth.currentUser;
+
   if (user) {
     // Si el usuario está autenticado, guardar el post
     addDoc(collection(db, "posts"), {
-      contenido: JSON.stringify(post), // Convertir a cadena JSON
+       ...post, 
       fecha: new Date(),
       usuario: user.uid,
-    });
+    })
+      .then((docRef) => {
+        console.log("Post guardado con ID: ", docRef.id);
+      })
+      .catch((error) => {
+        console.error("Error al guardar post: ", error);
+      });
   }
 };
 
 // Renderizar post
-export const getPosts = () => {
-  const postsQuery = query(collection(db, 'posts'), orderBy('fecha', 'desc'));
-  return getDocs(postsQuery)
-    .then((querySnapshot) => {
-      const posts = querySnapshot.docs.map((doc) => doc.data());
-      return posts;
-    })
-    .catch((error) => {
-      console.error('Error al cargar los posts: ', error);
-      return [];
-    });
-};
+// export const getPosts = () => {
+//   const postsQuery = query(collection(db, 'posts'), orderBy('fecha', 'desc'));
+//   return getDocs(postsQuery)
+//     .then((querySnapshot) => {
+//       const posts = querySnapshot.docs.map((doc) => doc.data());
+//       return posts;
+//     })
+//     .catch((error) => {
+//       console.error('Error al cargar los posts: ', error);
+//       return [];
+//     });
+// };
 
+export const getPosts = (callback) => {
+  const postsQuery = query(collection(db, 'posts'), orderBy('fecha', 'desc'));
+  return onSnapshot(postsQuery, callback)
+};
 
 // Eliminar post
 export const eliminarPost = (postId) => {
-  deleteDoc(doc(db, "posts", postId))
-    .then(() => {
-      console.log("Post eliminado con ID: ", postId);
-    })
-    .catch((error) => {
-      console.error("Error al eliminar post: ", error);
-    });
+  const confirmacion = window.confirm("¿Estás seguro de que deseas eliminar este post?");
+
+  if (confirmacion) {
+    deleteDoc(doc(db, "posts", postId))
+      .then(() => {
+        console.log("Post eliminado con ID: ", postId);
+      })
+      .catch((error) => {
+        console.error("Error al eliminar post: ", error);
+      });
+  }
 };
+
 
 // Editar post
 export const editarPost = (postId, nuevoContenido) => {
-  updateDoc(doc(db, "posts", postId), {
-    contenido: nuevoContenido,
+
+  const postRef = doc(db, "posts", postId);
+
+  updateDoc(postRef, {
+    contenido: nuevoContenido
   })
     .then(() => {
       console.log("Post editado con ID: ", postId);
